refactor: move start to seperate file
